Fixup the public_api test.
This was omitted in the rebase of the patch to split capi source
into a seprate crate. Move the public.rs test into its correct
location and move minimal.mp4 alongside it.

There's no great solution here, except perhaps using a different
test file for each crate. But I want the rust api to be the dominant
one in the long run, so prefer that location even though it crufts
up the capi documentation with a longer path.

We weren't packaging minimal.mp4 anyway because of the 'exclude'
line in Cargo.toml so this is not a regression; the tests dependent
on this file technically wouldn't pass from the published source.

Use the same exclude line in both crate manifests. Cargo now
warns about a dirty tree instead of packaging files which aren't
in version control, but this is still tedious with respect to
test files.